Cache the public directory (#51)
Looking through the docs and the logs, both the `.cache` and `public` directories need to be saved. If one is missing, the whole cache is ignored, hence our large build times. 
```
info We've detected that the Gatsby cache is incomplete (the .cache directory exists
but the public directory does not). As a precaution, we're deleting your site's
cache to ensure there's no stale data.
```

This change should improve future build times.